Move pusalink coordinators to separate module
