+++ homeassistant/components/prusalink/sensor.py
@@ -29,7 +29,9 @@ from homeassistant.helpers.typing import StateType
 from homeassistant.util.dt import utcnow
 from homeassistant.util.variance import ignore_variance
 
-from . import DOMAIN, PrusaLinkEntity, PrusaLinkUpdateCoordinator
+from . import PrusaLinkEntity
+from .const import DOMAIN
+from .coordinator import PrusaLinkUpdateCoordinator
 
 T = TypeVar("T", PrinterStatus, LegacyPrinterStatus, JobInfo)
